requirements.txt
albumentations==1.0.0
matplotlib==3.4.2
numpy==1.22.4
opencv_contrib_python==4.5.2.52
opencv_python==4.5.2.52
opencv_python_headless==4.5.2.52
pandas==2.0.3
Pillow==10.0.1
scikit_learn==1.2.1
timm==0.9.7
torch==2.0.1
tqdm==4.61.0
wandb==0.15.12
flake8==6.1.0
gradio==3.5.0
